requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.0.0
pydantic>=2.0.0
python-multipart==0.0.20
google-generativeai==0.8.5 
google-genai==1.15.0
sentence_transformers==4.1.0

pillow>=10.0.0
aiofiles>=24.1.0
asyncio-throttle>=1.0.2
openai>=1.82.0
groq>=0.25.0
httpx>=0.28.1
numpy>=2.2.6
torch>=2.7.0
transformers>=4.52.3